+++ mockserver/mockserver-core/src/main/java/org/mockserver/serialization/model/HttpResponseModifierDTO.java
@@ -1,5 +1,6 @@
 package org.mockserver.serialization.model;
 
+import com.fasterxml.jackson.databind.JsonNode;
 import org.mockserver.model.HttpResponseModifier;
 import org.mockserver.model.ObjectWithReflectiveEqualsHashCodeToString;
 
@@ -12,6 +13,8 @@ public class HttpResponseModifierDTO extends ObjectWithReflectiveEqualsHashCodeT
     private CookiesModifierDTO cookies;
     private HttpResponseModifierConditionDTO condition;
     private List<HttpResponseModifierDTO> modifiers;
+    private JsonNode jsonPatch;
+    private JsonNode jsonMergePatch;
 
     public HttpResponseModifierDTO() {
     }
@@ -24,6 +27,8 @@ public class HttpResponseModifierDTO extends ObjectWithReflectiveEqualsHashCodeT
             modifiers = httpResponseModifier.getModifiers() != null
                 ? httpResponseModifier.getModifiers().stream().map(HttpResponseModifierDTO::new).collect(Collectors.toList())
                 : null;
+            jsonPatch = httpResponseModifier.getJsonPatch();
+            jsonMergePatch = httpResponseModifier.getJsonMergePatch();
         }
     }
 
@@ -32,7 +37,9 @@ public class HttpResponseModifierDTO extends ObjectWithReflectiveEqualsHashCodeT
             .withHeaders(headers != null ? headers.buildObject() : null)
             .withCookies(cookies != null ? cookies.buildObject() : null)
             .withCondition(condition != null ? condition.buildObject() : null)
-            .withModifiers(modifiers != null ? modifiers.stream().map(HttpResponseModifierDTO::buildObject).collect(Collectors.toList()) : null);
+            .withModifiers(modifiers != null ? modifiers.stream().map(HttpResponseModifierDTO::buildObject).collect(Collectors.toList()) : null)
+            .withJsonPatch(jsonPatch)
+            .withJsonMergePatch(jsonMergePatch);
     }
 
     public HeadersModifierDTO getHeaders() {
@@ -68,6 +75,24 @@ public class HttpResponseModifierDTO extends ObjectWithReflectiveEqualsHashCodeT
 
     public HttpResponseModifierDTO setModifiers(List<HttpResponseModifierDTO> modifiers) {
         this.modifiers = modifiers;
+        return this;
+    }
+
+    public JsonNode getJsonPatch() {
+        return jsonPatch;
+    }
+
+    public HttpResponseModifierDTO setJsonPatch(JsonNode jsonPatch) {
+        this.jsonPatch = jsonPatch;
+        return this;
+    }
+
+    public JsonNode getJsonMergePatch() {
+        return jsonMergePatch;
+    }
+
+    public HttpResponseModifierDTO setJsonMergePatch(JsonNode jsonMergePatch) {
+        this.jsonMergePatch = jsonMergePatch;
         return this;
     }
 }